requirements.txt
pandas
numpy
matplotlib
seaborn
shap
scikit-learn
statsmodels
patsy
pyyaml
tabulate
psycopg2
sqlalchemy
